+++ feather/examples/async_connect_saved.rs
@@ -3,7 +3,7 @@
 
 use embassy_time::Timer;
 use feather::hal::ehal::digital::OutputPin;
-use feather::init3::init;
+use feather::init_async::init;
 use feather::shared::SpiStream;
 use wincwifi::AsyncClient;
 use wincwifi::StackError;
