+++ feather/examples/connect_network.rs
@@ -7,7 +7,7 @@
 use bsp::shared::SpiStream;
 use feather as bsp;
 use feather::hal::ehal::digital::OutputPin;
-use feather::init2::init;
+use feather::init::init;
 use feather::shared::{create_countdowns, delay_fn};
 
 const DEFAULT_TEST_SSID: &str = "network";
