+++ feather/examples/ping.rs
@@ -9,7 +9,7 @@ use core::net::Ipv4Addr;
 use bsp::shared::SpiStream;
 use feather as bsp;
 use feather::hal::ehal::digital::OutputPin;
-use feather::init2::init;
+use feather::init::init;
 use feather::shared::{create_countdowns, delay_fn, parse_ip_octets};
 
 use core::str::FromStr;
